fix(plm): teach the sega F1 tripwire about Peanuts, and close three Lows (#1217)
BLOCKING -- supabase/tests/sega_landing_contracts.sql F1 failed with 19
classification changes. Another author's file, changed deliberately and
minimally; nothing else in it was touched.

F1 asserts that api.source_capture_inventory still classifies every pre-existing
source exactly as before. My 19 plm.peanuts_* tables legitimately change that
set, so the tripwire was right to fire and the expectation was stale. The
expectation now derives from a v_map array declared once at the top of the
block, in the same order the view tests its prefixes, and still written
INDEPENDENTLY of the view -- which is what keeps it a tripwire rather than a
tautology. A pre-existing source that changes classification still fails with
the regression message.

What is new is that an unknown prefix no longer passes silently. A table the map
has never heard of, which the view classifies as something, now fails with a
message naming v_map and saying to add one row. That turns the #1195 phase-2
tripwire failure into a one-line fix with instructions, instead of a puzzle. My
own section G was restructured to the identical shape so the two files agree.

Verified by mutation: deleting the pre-existing nbcu arm from the view still
fails F1 with "plm.nbcu_capture is now other but was nbcu"; adding an unknown
source arm fails with the add-a-row message; removing the peanuts row from
v_map fails, so the fix is not vacuous.

LOWS
* The zero-media guard family was untested. C6 now proves the hard-zero CHECK
  refuses a media-bearing capture by name, and asserts structurally that the
  completion rule still carries its media clause -- that clause is unreachable
  by insert while the column CHECK stands, and asserting a defence-in-depth
  clause is still there is the honest way to pin it. F2 gains the
  {"media_downloaded":1} manifest case, which is the ONLY check anywhere on a
  manifest claiming media, since the finalize sweep would accept it as a
  well-formed extra key. finalize_'s media_downloaded_nonzero branch is
  genuinely unreachable while the column CHECK stands; it is now commented as
  deliberate defence in depth rather than reported as covered.
* begin_'s out-of-range branch had no failing test. F2 gains an oversized shape.
* The completion CHECK summed two `integer` columns uncast, so an owner row
  above 2^31-1 raised raw `integer out of range` instead of check_violation.
  All three terms cast to bigint, matching the function, with test C7.

Found while mutating: C6 and C7 as first written only raised a WARNING when the
bad row was ACCEPTED, so dropping peanuts_capture_media_zero_chk left the suite
green. Both now set an explicit fired flag and raise. Every other warning path
in the file sits behind a counter that is asserted; audited, all clear.

Refs #1217. Claim #1231.

+++ supabase/migrations/20260819125713_peanuts_tenovos_source_landing.sql
@@ -167,6 +167,11 @@ create table plm.peanuts_capture (
   constraint peanuts_capture_manifest_sha256_chk
     check (source_manifest_sha256 ~ '^[0-9a-f]{64}$'),
   -- Scope guard, not a style preference: this project may never download Peanuts media.
+  -- This is the constraint that makes the media clause of
+  -- peanuts_capture_complete_requirements_chk below, and the media_downloaded_nonzero
+  -- branch of plm.finalize_peanuts_capture, UNREACHABLE in normal operation. Both are
+  -- kept deliberately as defence in depth for a future in which this line is dropped;
+  -- see the note on that branch in the function.
   constraint peanuts_capture_media_zero_chk check (media_downloaded = 0),
   constraint peanuts_capture_portal_total_nonneg_chk
     check (portal_reported_asset_total >= 0),
@@ -187,7 +192,15 @@ create table plm.peanuts_capture (
         and vocabularies_loaded_from_source = true
         and media_downloaded = 0
         and jsonb_array_length(error_summary) = 0
-        and assets_captured + assets_unreachable = portal_reported_asset_total
+        -- CAST BEFORE ADDING. These three columns are `integer`; a bare
+        -- `assets_captured + assets_unreachable` raises a raw `integer out of range`
+        -- instead of a check_violation when an owner writes a row above 2^31-1, so the
+        -- caller sees a PostgreSQL arithmetic error rather than this named constraint.
+        -- Unreachable through any granted role, and plm.finalize_peanuts_capture already
+        -- casts on its own copy of this arithmetic -- the CHECK must match the function
+        -- rather than differ from it in the one place nobody looks.
+        and assets_captured::bigint + assets_unreachable::bigint
+            = portal_reported_asset_total::bigint
       )
     ),
   constraint peanuts_capture_expected_counts_obj_chk
@@ -1364,6 +1377,12 @@ begin
   if not v_cap.vocabularies_loaded_from_source then
     v_err := v_err || jsonb_build_object('code','vocabularies_not_loaded_from_source');
   end if;
+  -- UNREACHABLE WHILE peanuts_capture_media_zero_chk STANDS, AND KEPT ON PURPOSE.
+  -- The column is `not null default 0` with a CHECK pinning it to zero, so no writer can
+  -- put a nonzero value here today and no contract test can reach this branch without
+  -- first dropping that constraint. It is defence in depth for the day someone does.
+  -- Stated plainly rather than listed as tested, because a branch reported as covered
+  -- when it is not is how the next reader is misled.
   if v_cap.media_downloaded <> 0 then
     v_err := v_err || jsonb_build_object('code','media_downloaded_nonzero',
                                          'observed', v_cap.media_downloaded);
